refactor(authentication): move psql repo out of subdirectory
