+++ authentication/accounts/repository/ddl.go
@@ -1,4 +1,4 @@
-package postgresql
+package repository
 
 import _ "embed"
