+++ authentication/accounts/repository/sqlc.yaml
@@ -5,5 +5,5 @@ sql:
     schema: "schema.sql"
     gen:
       go:
-        package: "postgresql"
-        out: "generated"
+        package: "sqlc"
+        out: "_sqlc"
